pyproject.toml
[project]
name = "fastapi-base"
version = "0.1.0"
description = ""
authors = [
    {name = "Cesar Freire",email = "iceesar@live.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.1"
pytest-asyncio = "^1.0.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.2"
testcontainers = "^4.10.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', 'seed.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.lint.pylint]
max-nested-blocks = 10
max-statements = 75
max-args = 8
max-branches = 17

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fastapi_base/app.py'
run_host = 'fastapi dev fastapi_base/app.py --host 0.0.0.0'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastapi_base -vv'
post_test = 'coverage html'